Better separation of concerns
